Re introduce plugin.py file
